商品マスタ

+++ db/migrate/20240623043939_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration[7.1]
     create_table :products do |t|
       t.references :maker, null: false, foreign_key: true
       t.string :name
-      t.integer :jancode
+      t.bigint :jancode
       t.string :product_code
       t.decimal :price
       t.integer :unit
